requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fonttools==4.39.2
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.2
packaging==23.0
Pillow==9.4.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
response==0.5.0
scipy==1.10.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.15
